Fixed bug where external traffic policies would be erroneously deleted when intents from multiple namespaces existed and all intents in one namespace were deleted.

+++ src/operator/controllers/intents_reconcilers/network_policy.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 	otterizev1alpha2 "github.com/otterize/intents-operator/src/operator/api/v1alpha2"
-	"github.com/otterize/intents-operator/src/operator/controllers/external_traffic"
 	"github.com/otterize/intents-operator/src/shared/injectablerecorder"
 	"github.com/otterize/intents-operator/src/shared/telemetries/telemetriesgql"
 	"github.com/otterize/intents-operator/src/shared/telemetries/telemetrysender"
@@ -21,6 +20,7 @@ import (
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
+	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 )
 
 const (
@@ -36,7 +36,7 @@ const (
 type NetworkPolicyReconciler struct {
 	client.Client
 	Scheme                      *runtime.Scheme
-	endpointsReconciler         external_traffic.EndpointsReconcilerInterface
+	endpointsReconciler         reconcile.Reconciler
 	RestrictToNamespaces        []string
 	enableNetworkPolicyCreation bool
 	enforcementEnabledGlobally  bool
@@ -46,7 +46,7 @@ type NetworkPolicyReconciler struct {
 func NewNetworkPolicyReconciler(
 	c client.Client,
 	s *runtime.Scheme,
-	endpointsReconciler external_traffic.EndpointsReconcilerInterface,
+	endpointsReconciler reconcile.Reconciler,
 	restrictToNamespaces []string,
 	enableNetworkPolicyCreation bool,
 	enforcementEnabledGlobally bool) *NetworkPolicyReconciler {
@@ -254,7 +254,7 @@ func (r *NetworkPolicyReconciler) handleNetworkPolicyRemoval(
 	var intentsList otterizev1alpha2.ClientIntentsList
 	err := r.List(
 		ctx, &intentsList,
-		&client.MatchingFields{otterizev1alpha2.OtterizeTargetServerIndexField: intent.Name},
+		&client.MatchingFields{otterizev1alpha2.OtterizeTargetServerIndexField: intent.GetServerFullyQualifiedName(intentsObjNamespace)},
 		&client.ListOptions{Namespace: intentsObjNamespace})
 
 	if err != nil {
@@ -289,20 +289,39 @@ func (r *NetworkPolicyReconciler) deleteNetworkPolicy(
 		return err
 	}
 
+	// Check if network policies from all namespaces are gone
+	// If so, check if intents required flag for external policies is disabled
+	// If so, delete.
+
 	// Remove network policies created by the external traffic reconcilers.
 	// Once no more Otterize network policies are present, there's no longer need for them.
-	externalPolicyList := &v1.NetworkPolicyList{}
-	serviceNameLabel := policy.Labels[otterizev1alpha2.OtterizeNetworkPolicy]
-	err = r.List(ctx, externalPolicyList, client.MatchingLabels{otterizev1alpha2.OtterizeNetworkPolicyExternalTraffic: serviceNameLabel},
-		&client.ListOptions{Namespace: policy.Namespace})
+	// Before we do this, we must check there are no Otterize network policies in ANY namespace.
+	// This function is called once there are no more network policies from this namespace.
+
+	var intentsList otterizev1alpha2.ClientIntentsList
+	err = r.List(
+		ctx, &intentsList,
+		&client.MatchingFields{otterizev1alpha2.OtterizeTargetServerIndexField: intent.GetServerFullyQualifiedName(intentsObjNamespace)})
+
 	if err != nil {
 		return err
 	}
 
-	for _, externalPolicy := range externalPolicyList.Items {
-		err := r.Delete(ctx, externalPolicy.DeepCopy())
+	// This is the last intent out of all namespaces, so it's safe to delete the external traffic policy.
+	if len(intentsList.Items) == 1 {
+		externalPolicyList := &v1.NetworkPolicyList{}
+		serviceNameLabel := policy.Labels[otterizev1alpha2.OtterizeNetworkPolicy]
+		err = r.List(ctx, externalPolicyList, client.MatchingLabels{otterizev1alpha2.OtterizeNetworkPolicyExternalTraffic: serviceNameLabel},
+			&client.ListOptions{Namespace: policy.Namespace})
 		if err != nil {
 			return err
+		}
+
+		for _, externalPolicy := range externalPolicyList.Items {
+			err := r.Delete(ctx, externalPolicy.DeepCopy())
+			if err != nil {
+				return err
+			}
 		}
 	}
